make the pidfile optional
Normally bitmarkd is run with the aid of daemon(8) of FreeBSD or
systemd on Linux or some similar supervisor system on other OS.
Therefore there is no need to create a pidfile, hoever the "pidfile =
" configuration directive can be used to create one.

If the pidfile configuration is used, bitmarkd will not detect a stale
pidfile, it simply refuses to start; manually delete the pidfile in
this case.

Signed-off-by: Christopher Hall <hsw@bitmark.com>