[block,fault,mode,payment,peer,rpc,storage,transactionrecord] basic transfer operation
This make transfers basically work and update the ownership data.

However there is a problem caused by the block delete method as this
can result in a chain of transfers or double transfers being returned
to the verified tables.  Which will case a crash as soon as the now
invalid transfer is encountered.  The crash will show as corrupt
database, but it really is not as the once verified transfer should
really be dropped.

commit just to save the current state before removing the verified
tables and consolidating them into the current memory cache.

Signed-off-by: Christopher Hall <hsw@bitmark.com>