addition of batch and fingerprint commands
batch is a testing operation that creates a number of issues, creates
corresponding number of new accounts and transfers one newly issued
bitmark to each account.  Finally displays the result as a JSON object
allowing a scripting with "jq" to extract the accounts, bitmark-pay
command etc. for further processing.

fingerprint is to print the hex fingerprint of a file and produces the
same result as the desktop app to allow for compatible issue
operations.

also various refactoring was carried out.

Signed-off-by: Christopher Hall <hsw@bitmark.com>